Refactor beam_job_domain file
