+++ core/domain/beam_job_domain.py
@@ -25,17 +25,6 @@ from core.jobs import base_jobs
 
 from typing import Dict, List, Type, Union
 
-from core.platform import models  # pylint: disable=invalid-import-from # isort:skip
-
-# TODO(#14537): Refactor this file and remove imports marked
-# with 'invalid-import-from'.
-
-MYPY = False
-if MYPY:  # pragma: no cover
-    from mypy_imports import beam_job_models
-
-(beam_job_models,) = models.Registry.import_models([models.NAMES.beam_job])
-
 
 class BeamJob:
     """Encapsulates the definition of an Apache Beam job.
@@ -121,22 +110,6 @@ class BeamJobRun:
         self.job_started_on = job_started_on
         self.job_updated_on = job_updated_on
         self.job_is_synchronous = job_is_synchronous
-
-    @property
-    def in_terminal_state(self) -> bool:
-        """Returns whether the job run has reached a terminal state and is no
-        longer executing.
-
-        Returns:
-            bool. Whether the job has reached a terminal state.
-        """
-        return self.job_state in (
-            beam_job_models.BeamJobState.CANCELLED.value,
-            beam_job_models.BeamJobState.DRAINED.value,
-            beam_job_models.BeamJobState.UPDATED.value,
-            beam_job_models.BeamJobState.DONE.value,
-            beam_job_models.BeamJobState.FAILED.value,
-        )
 
     def to_dict(self) -> Dict[str, Union[bool, float, str, List[str]]]:
         """Returns a dict representation of the BeamJobRun.
